""ClusterService"" to swagger.json
